🚀 #55 - split single job into batch and utilise full power of several workers

+++ app/Jobs/GithubJob.php
@@ -22,13 +22,19 @@ abstract class GithubJob extends Job implements ShouldBeUnique
     public function __construct()
     {
         $this->queue = 'github';
-        $this->timeout = CarbonInterval::hour()->totalSeconds;
+        $this->timeout = CarbonInterval::minute()->totalSeconds;
     }
 
-    public function handle(): bool
+    public function handle(): ?bool
     {
+        if ($this->batch() !== null && $this->batch()->cancelled()) {
+            return null;
+        }
+
         if ($this->entity()->isBlocked()) {
             $this->delete();
+
+            return null;
         }
 
         try {
@@ -96,7 +102,12 @@ abstract class GithubJob extends Job implements ShouldBeUnique
     public function backoff(): int | array
     {
         return [
+            CarbonInterval::seconds(5)->totalSeconds,
+            CarbonInterval::seconds(10)->totalSeconds,
+            CarbonInterval::seconds(30)->totalSeconds,
             CarbonInterval::minute()->totalSeconds,
+            CarbonInterval::minutes(5)->totalSeconds,
+            CarbonInterval::minutes(10)->totalSeconds,
             CarbonInterval::minutes(15)->totalSeconds,
             CarbonInterval::minutes(30)->totalSeconds,
             CarbonInterval::hour()->totalSeconds,
